Extract ZSlider into its own module
