+++ mantidimaging/gui/widgets/zslider/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2023 ISIS Rutherford Appleton Laboratory UKRI
+# SPDX - License - Identifier: GPL-3.0-or-later
+from __future__ import annotations
